Divide code into multiple files
